Automatically adjust font-size and use unicode control characters to manage line breaks.
Spaces are replaced with non-breakable spaces, to disable word wrap in
general. Then, zero-width spaces are inserted in some precise locations
to allow word wrap there, but only when needed (Qt manages line breaks
and makes nice lines of the proper length :-) ).

In addition to that, QFontMetrics is used to measure the rendered lines,
and progressively make the font smaller until all the lines fit on the
screen. So, in case a line is really impossible to break properly, we
still ensure that the whole line is shown on screen.